test: test functions for exif edit
refactoring the code for test functions for exifedit
